+++ api/highlight_server/highlight_server/asgi.py
@@ -1,0 +1,16 @@
+"""
+ASGI config for highlight_server project.
+
+It exposes the ASGI callable as a module-level variable named ``application``.
+
+For more information on this file, see
+https://docs.djangoproject.com/en/3.0/howto/deployment/asgi/
+"""
+
+import os
+
+from django.core.asgi import get_asgi_application
+
+os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'highlight_server.settings')
+
+application = get_asgi_application()
