+++ api/highlight_server/server/all_errors.txt
@@ -1,0 +1,13 @@
+"""
+responses:
+:errors 1000: can't register user
+:errors 2000: no account matching login
+:errors 2001: wrong password
+:errors 2002: account not verified
+:errors 2003: no account matching id
+:errors 2004: account has not enough rights
+:errors 3000: pieces already taken
+:errors 3001: not a single piece (как у вас могут быть карты не в том порядке разложены?!)
+:errors 4040: a error
+:success OK: no error occurred during operation
+"""
