+++ api/highlight_server/server/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class ServerConfig(AppConfig):
+    name = 'server'
